test: extract shared proxy-env fixture
The proxy env-var snapshot/clear/restore scaffolding was duplicated
verbatim in the fetch-with-retry and http-dispatcher transport tests.
Move it to src/_fixtures/proxy-env.ts so both suites share one
implementation and can't drift on teardown.
