+++ src/transport/fetch-with-retry.test.ts
@@ -1,33 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-
-const PROXY_ENV_KEYS = [
-    'HTTP_PROXY',
-    'http_proxy',
-    'HTTPS_PROXY',
-    'https_proxy',
-    'NO_PROXY',
-    'no_proxy',
-] as const
-
-const originalProxyEnv = new Map(PROXY_ENV_KEYS.map((key) => [key, process.env[key]]))
-
-function clearProxyEnv(): void {
-    for (const key of PROXY_ENV_KEYS) {
-        delete process.env[key]
-    }
-}
-
-function restoreProxyEnv(): void {
-    for (const key of PROXY_ENV_KEYS) {
-        const value = originalProxyEnv.get(key)
-        if (value === undefined) {
-            delete process.env[key]
-            continue
-        }
-
-        process.env[key] = value
-    }
-}
+import { clearProxyEnv, restoreProxyEnv } from '../_fixtures/proxy-env.js'
 
 describe('fetchWithRetry', () => {
     beforeEach(() => {
